requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
cycler==0.10.0
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
joblib==0.13.2
Keras==2.1.4
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
nltk==3.4.4
numpy==1.14.5
pandas==0.25.0
protobuf==3.9.1
pydot==1.4.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.15.5
